apps/api/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
websockets>=13.0
mcp>=1.3.0
python-dotenv>=1.0.0
